requirements.txt
Flask>=2.2
pdfminer.six>=20221105
PyPDF2>=3.0
openai>=1.0.0
beautifulsoup4>=4.12
nltk>=3.8
requests>=2.31
python-dotenv>=1.0.0
gunicorn>=20.1.0
scikit-learn>=1.3
pytest>=7.0